libs/probill/probill/requirements.txt
aiohttp==3.9.3
openai==1.12.0
aiofiles==23.2.1
bcrypt==4.1.1
cryptography==41.0.7
Pillow==10.1.0
scikit-learn==1.3.2
fastdtw==0.3.4
matplotlib==3.8.2
pycryptodome==3.20.0
dspy-ai[groq]
chromadb == 0.5.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
psycopg2-binary==2.9.9
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langgraph==0.1.1
langsmith==0.1.75
Pydantic==2.5.2
python-ulid==2.7.0
aiosqlite==0.20.0
pyotp==2.9.0
PyNaCl==1.5.0
cffi==1.17.0
ujson==5.10.0
mypy==1.11.2